transport_system/backend/requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-redis>=5.0.0
channels>=4.0.0
channels-redis>=4.0.0
celery>=5.3.0
redis>=4.5.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
Pillow>=10.0.0
stripe>=6.0.0


